Split machine into its own module
